refactor: separate utils
